Better error handling for CETMM core

+++ scripts/auth.lua
@@ -1,16 +1,16 @@
 local auth = {}
 
 function auth.GetCore()
-  json.CETMM_core(0)
-  local file = io.open("scripts/hash", "r")
-  if file then
-    local hash = file:read()
-    file:close()
-    os.remove("scripts/hash")
-    return json.CETMM_core(hash)
-  else
-    error("can't access to CETMM core.", 2)
+  if pcall(json.CETMM_core, 0) then
+    local file = io.open("scripts/hash", "r")
+    if file then
+      local hash = file:read()
+      file:close()
+      os.remove("scripts/hash")
+      return json.CETMM_core(hash)
+    end
   end
+  error("can't access CETMM core library. Check if cet_mod_manager.asi loaded.", 1)
 end
 
 return auth
